requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
exchangelib>=5.2.0
psycopg2-binary>=2.9.9
requests>=2.31.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
pytest>=7.4.3
cryptography>=42.0.0
PyMuPDF>=1.23.0
python-docx>=1.1.0
sentence-transformers>=2.3.0
